+++ src/textual/layouts/horizontal.py
@@ -81,7 +81,7 @@ class HorizontalLayout(Layout):
             children, box_models, margins
         ):
             styles = widget.styles
-            has_rule = styles.has_rule
+
             overlay = styles.overlay == "screen"
             offset = (
                 styles.offset.resolve(
@@ -109,9 +109,12 @@ class HorizontalLayout(Layout):
                 overlay,
             )
 
-            if has_rule("constrain_x") or has_rule("constrain_y"):
-                placement = placement.apply_constrain(
-                    viewport.region if placement.overlay else size.region
+            if (
+                styles.has_any_rules("constrain_x", "constrain_y")
+                or widget.absolute_offset is not None
+            ):
+                placement = placement.process_offset(
+                    viewport.region if overlay else size.region
                 )
             add_placement(placement)
             if not overlay: